requirements.txt
Flask==2.2.4
Flask-SQLAlchemy==2.5.1
flask-cors==3.0.10
SQLAlchemy==1.4.47
psycopg2-binary==2.9.3
Werkzeug==2.3.7
python-dotenv==1.0.0
